feat: add new pages and refactor the routing
